+++ mokkoji_backend/src/main/java/com/example/mokkoji_backend/service/myPage/CartService.java
@@ -14,7 +14,7 @@ public interface CartService {
 	// 2) cart의 총 개수 조회
 	int countCart(String userId);
 
-	void updateCart(String userId, int productId, String optionContent, String packagingOptionContent, int productCnt, int productTotalCount);
+	void updateCart(String userId, long productId, String optionContent, String packagingOptionContent, int productCnt, int productTotalCount);
 
 	void deleteCart(CartId cartId);
